requirements.txt
anthropic==0.34.2
fastapi==0.115.0
motor==3.5.1
pip-chill==1.0.3
python-dotenv==1.0.1
